refactor(shared-server): remove group mutation compatibility alias
