+++ packages/shared-server/rallar-system/group-state/mutation/result-validation/validate-computed-group-mutation.ts
@@ -8,19 +8,19 @@ import type {
 import { validateComputedGroupMutationWrite } from './validate-computed-group-mutation-write.ts';
 import { validateCommandHash, validateMutationReceipt } from './validate-group-mutation-result.ts';
 
-export interface ValidateComputedMutationShapeInput {
+export interface ValidateComputedGroupMutationInput {
     readonly command: GroupMutationCommand;
     readonly read: GroupMutationRead;
     readonly facts: GroupMutationFacts;
     readonly computed: GroupMutationComputed;
 }
 
-export function validateComputedMutationShape({
+export function validateComputedGroupMutation({
     command,
     read,
     facts,
     computed
-}: ValidateComputedMutationShapeInput): void {
+}: ValidateComputedGroupMutationInput): void {
     const value = computed;
     switch (computed.outcome) {
         case 'replay':
@@ -116,6 +116,3 @@ function validateWriteOutcomeKeys(value: object): void {
     assertExactKeys(value, keys, 'Group mutation computed result');
     assertRequiredKeys(value, keys, 'Group mutation computed result');
 }
-
-export { validateComputedMutationShape as validateComputedGroupMutation };
-export { validateComputedRosterFacts } from '../state-validation/validate-computed-roster-facts.ts';
