+++ packages/shared-server/rallar-system/group-state/mutation/state-validation/validate-group-mutation.ts
@@ -11,11 +11,9 @@ import type {
     GroupMutationRead
 } from '../group-mutation-contracts.ts';
 import { computeGroupMutation } from '../orchestration/compute-group-mutation.ts';
-import {
-    validateComputedMutationShape,
-    validateComputedRosterFacts
-} from '../result-validation/validate-computed-group-mutation.ts';
+import { validateComputedGroupMutation } from '../result-validation/validate-computed-group-mutation.ts';
 import { validateComputedGroupMutationOutbox } from '../result-validation/validate-computed-group-mutation-outbox.ts';
+import { validateComputedRosterFacts } from './validate-computed-roster-facts.ts';
 import { validateGroupMutationFacts } from './validate-group-mutation-facts.ts';
 import { validateGroupMutationRead } from './validate-group-mutation-read.ts';
 
@@ -35,7 +33,7 @@ export function validateGroupMutation(
         input.computed.outcome === 'write' ? { ...input.computed, outboxEntries: [] } : input.computed,
         'Group mutation computed result'
     );
-    validateComputedMutationShape(input);
+    validateComputedGroupMutation(input);
     const canonical = computeGroupMutation({
         command: input.command,
         read: input.read,
